Added a size component, a static collision class and fleshed out the World class
The size component, along with the already existing position component,
allows for interaction between Tiles and GameActors despite them being
fundamentally different (at least for now).

The World class has been given an initial skeleton that accepts a
tileset, map size and a source (tile representation in a text file,
likely doesn't need the map size because of this actually) and builds a
map with fully described Tiles. The implementation is somewhat jagged
and needs refining.

Lastly, added a static collision class that uses SizeComponents and
PositionComponents as opposed to GameActors, this allows GameActors to
interact identically with Tiles.