Streamlined the World class and made it functional
Streamlined the World class so that all variables are created in one
place and the Class is initialised from the constructor rather than
setting values after an empty constructor.
The class now works and should be able to read a file in the appropriate
format although it still requires cleaning up.
One or two variables may be able to be removed from the World
constructor as they can be interpreted from the text file that it will
read.