Updated movement to use velocity
Movement now uses velocity and added full collision for two objects
whether moving or not. The World class now interpolates the movement of
the previous turn based on the respective velocites and angle of
collision for the two collided objects. Will likely bug if velocity is
too high, but shouldn't be a problem here. Fixable by extrapolating
relative position from the previous turn instead.