Added Conversation
Added a conversation component for the gameActors that initiates a
conversation in the World class. A field is created and passed back to
the World using the observer pattern and the world establishes a
Conversation class with the nearest gameActor and initiates it for a 3
second duration. This is begun with a Talk command.

Also added DialogueTree and Quest classes as a vestige for allowing
quests and dialogue to be initiated through conversations.